Add(user_dashboard-apis) : added user dashboard apis
